Moves OrderSummary into Contracts.
